moved styled tabs to another file, added tabcontents to sidebar
