+++ client/src/components/sidebar/sidebar.component.jsx
@@ -1,36 +1,24 @@
 import React, { useState } from 'react';
-import { Box, Grid, Typography, Tabs, Tab, makeStyles, withStyles } from '@material-ui/core';
+import { Box, Grid, Typography, makeStyles, withStyles } from '@material-ui/core';
 
 import Profile from '../../components/profile/profile.component';
+import { default as Tabs } from '../../components/tabs/tabs.component';
 
-const StyledTabs = withStyles({
-  indicator: {
-    display: 'flex',
-    justifyContent: 'center',
-    backgroundColor: 'transparent',
-    '& > div': {
-      maxWidth: 40,
-      width: '100%',
-      backgroundColor: '#635ee7',
-    },
-  },
-})(props => <Tabs {...props} TabIndicatorProps={{ children: <div /> }} />);
+function TabPanel(props) {
+  const { children, value, index, ...other } = props;
 
-const StyledTab = withStyles(theme => ({
-  root: {
-    textTransform: 'none',
-    color: '#000',
-    fontWeight: theme.typography.fontWeightRegular,
-    //    fontSize: theme.typography.pxToRem(18),
-    maxWidth: 100,
-    minWidth: 90,
-    marginRight: theme.spacing(1),
-    padding: 0,
-    '&:focus': {
-      opacity: 1,
-    },
-  },
-}))(props => <Tab disableRipple {...props} />);
+  return (
+    <Typography
+      component="div"
+      role="tabpanel"
+      hidden={value !== index}
+      id={`simple-tabpanel-${index}`}
+      {...other}
+    >
+      {value === index && <Box p={3}>{children}</Box>}
+    </Typography>
+  );
+}
 
 const Sidebar = () => {
   const [tab, setTab] = useState('Chats');
@@ -45,11 +33,18 @@ const Sidebar = () => {
         <Profile />
       </Grid>
       <Grid item>
-        <StyledTabs value={tab} onChange={changeTab} aria-label="styled tabs">
-          <StyledTab value="Chats" label="Chats" />
-          <StyledTab value="Contacts" label="Contacts" />
-          <StyledTab value="Invites" label="Invites" />
-        </StyledTabs>
+        <Tabs value={tab} onChange={changeTab}></Tabs>
+      </Grid>
+      <Grid item>
+        <TabPanel value={tab} index={'Chats'}>
+          Chats
+        </TabPanel>
+        <TabPanel value={tab} index={'Contacts'}>
+          Contacts
+        </TabPanel>
+        <TabPanel value={tab} index={'Invites'}>
+          Invites
+        </TabPanel>
       </Grid>
     </Box>
   );
